cleanup old lexer in semantic
